Worked on isCorrect() and checkAnswer()
Almost there, but still fuzzy on when datesThisRound should be created.

I think at the time of checking, but still not sure whether inside isCorrect() or checkAnswer. The result of datesThisRound is what needs to be passed into isCorrect, so I think it needs to be created outside. Maybe there's a separate getDatesThisRound() function that's called in checkAnswer?