Timer almost working
Reverted back to making timer a property of viewController because I was troubleshooting something else. Fixed it, so eventually I could switch back to making it a property of myGameManager. However, the main issue now is that rounds 2-6 display the last known value of the timer for the first second of their round, then continue down from :59 as expected. Have tried moving my timer reset message all around to no avail.