Fixed timer display
Manually set timer display in viewDidLoad() so it correctly displays for the first second after reloading.
Erik said: As for the next round not starting at 60 seconds, you should go ahead and assign the label text to 60 at the start of the next round like you do in your counter function.

The timer doesn't call the `counter` method when you start it. It only starts calling it after a second, which is why you have to set the text to 60 when you start the timer.